requirements.txt
absl-py==1.4.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.4
comm==0.1.2
contourpy==1.1.0
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
idna==2.10
importlib-metadata==6.0.0
importlib-resources==5.2.0
ipykernel==6.19.2
ipython==8.12.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.1.2
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==0.8.4
nb-conda==2.2.1
nb-conda-kernels==2.3.1
nbclassic==0.5.5
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.0.74
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==23.1.2
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roboflow==1.1.0
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
sentry-sdk==1.27.1
setproctitle==1.3.2
setuptools==67.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.4
stack-data==0.2.0
supervision==0.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
terminado==0.17.1
thop==0.1.1.post2209072238
tinycss2==1.2.1
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.3.6
wget==3.2
wheel==0.38.4
zipp==3.11.0
